+++ inst/doc/ggblanket.R
@@ -16,6 +16,7 @@ library(ggplot2)
 library(dplyr)
 library(stringr)
 library(palmerpenguins)
+library(patchwork)
 
 ## -----------------------------------------------------------------------------
 iris |>
@@ -48,8 +49,7 @@ penguins |>
     x = flipper_length_mm,
     col = sex,
     facet = species,
-    facet2 = island
-  )
+    facet2 = island)
 
 ## -----------------------------------------------------------------------------
 penguins |> 
@@ -73,13 +73,13 @@ penguins |>
     y_include = 0,
     y_trans = "sqrt",
     y_breaks = scales::breaks_width(1500), 
-    y_labels = scales::label_number()
-  )
+    y_labels = scales::label_number())
 
 ## -----------------------------------------------------------------------------
 penguins |>
   group_by(species, sex) |>
-  summarise(across(flipper_length_mm, ~ round(mean(.x, na.rm = TRUE)), 0)) |> 
+  summarise(
+    flipper_length_mm = round(mean(flipper_length_mm, na.rm = TRUE), 0)) |>
   gg_tile(
     x = sex, 
     y = species, 
@@ -110,8 +110,7 @@ storms |>
     title = "Storm wind speed",
     subtitle = "USA average storm wind speed, 1980\u20132020",
     y_title = "Wind speed (knots)",
-    caption = "Source: NOAA"
-  ) +
+    caption = "Source: NOAA") +
   geom_point()
 
 ## -----------------------------------------------------------------------------
@@ -144,12 +143,12 @@ penguins |>
     x = flipper_length_mm,
     y = body_mass_g,
     col = sex,
-    size = 0.5, 
-    level = 0.99, 
+    linewidth = 0.5, #accessed via geom_smooth
+    level = 0.99, #accessed via geom_smooth
     col_legend_place = "t",
     col_title = "", 
-    col_labels = snakecase::to_sentence_case
-  ) 
+    col_labels = snakecase::to_sentence_case, 
+    colour = "white") #accessed via geom_smooth
 
 ## -----------------------------------------------------------------------------
 penguins |>
@@ -171,13 +170,93 @@ storms |>
     y = wind,
     x_labels = scales::label_comma(big.mark = ""),
     x_expand = c(0, 0),
-    width = 0.75,
     theme = gg_theme(
       text_size = 11,
       plot_background_pal = "white",
       panel_background_pal = "white"))
 
 ## -----------------------------------------------------------------------------
+nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
+
+nc |>
+  gg_sf(col = AREA,
+        title = "**Bold** or _italics_ or <span style = 'color:red;'>red</span>",
+        theme = gg_theme(title_face = "plain")) +
+  theme(plot.title = ggtext::element_markdown())
+
+## -----------------------------------------------------------------------------
+p1 <- diamonds |> 
+  gg_point(
+    x = carat,
+    y = price)
+
+p2 <- diamonds |> 
+  gg_bin2d(
+    x = carat,
+    y = price,
+    caption = "Source: Diamonds Association",
+    theme = gg_theme(caption_hjust = 1))
+
+p1 + p2
+
+## ---- eval = FALSE------------------------------------------------------------
+#  df <- iris |>
+#    tibble::tibble() |>
+#    add_tooltip(Sepal.Width, Sepal.Length, Species)
+#  
+#  df |>
+#    head(1)
+#  
+#  p <- iris |>
+#    add_tooltip(tidyselect::contains("Sepal"), Species) |>
+#    mutate(id = row_number()) |>
+#    gg_blank(x = Sepal.Width,
+#             y = Sepal.Length,
+#             col = Species,
+#             facet = Species,
+#             theme = gg_theme("helvetica"),
+#             facet_labels = snakecase::to_sentence_case) +
+#    ggiraph::geom_point_interactive(aes(tooltip = tooltip, data_id = id))
+#  
+#  ggiraph::girafe(
+#    ggobj = p,
+#    width_svg = 5,
+#    height_svg = 3.25,
+#    options = list(
+#      ggiraph::opts_tooltip(use_fill = TRUE, use_stroke = TRUE),
+#      ggiraph::opts_hover(css = "fill: #d62728; stroke: #d62728"))
+#  )
+
+## ---- echo = FALSE------------------------------------------------------------
+knitr::include_graphics("../man/figures/ggiraph_screenshot.png", dpi = 300)
+
+## -----------------------------------------------------------------------------
+p1 <- penguins |>
+  gg_blank(
+    x = flipper_length_mm,
+    y = body_mass_g,
+    col = species,
+    pal = rep(pal_viridis_mix(1), 3), 
+    x_breaks = scales::breaks_pretty(3),
+    col_legend_ncol = 2) +
+  geom_point(aes(alpha = species)) +
+  labs(alpha = "Species") +
+  scale_alpha_manual(values = c(0.1, 1, 0.5))
+
+p2 <- penguins |>
+  gg_blank(
+    x = flipper_length_mm, 
+    y = body_mass_g, 
+    col = species, 
+    x_breaks = scales::breaks_pretty(3), 
+    col_legend_ncol = 2) +
+  geom_point(aes(alpha = species)) +
+  labs(alpha = "Species") +
+  scale_alpha_manual(values = c(0.1, 1, 0.5))
+
+p1 + p2
+
+## ---- fig.asp=0.7-------------------------------------------------------------
 gg_point_custom <- function(data, x, y, col, 
                             size = 3, 
                             shape = 17,
@@ -203,94 +282,43 @@ iris |>
     col = Species, 
     title = "Edgar Anderson's iris data",
     subtitle = "Iris sepal length by width and species",
-    caption = "Edgar Anderson, 1935"
-  )
-
-## ---- eval = F----------------------------------------------------------------
-#  theme_custom <- gg_theme(
-#    "helvetica",
-#    plot_background_pal = "white",
-#    panel_background_pal = "white",
-#  )
-#  
-#  iris |>
-#    mutate(Species = str_to_sentence(Species)) |>
-#    add_tooltip_text(titles = snakecase::to_sentence_case) |>
-#    gg_point(
-#      x = Sepal.Width,
-#      y = Sepal.Length,
-#      col = Species,
-#      text = text,
-#      col_legend_place = "r",
-#      theme = theme_custom) |>
-#    plotly::ggplotly(tooltip = "text")
-
-## ---- echo=FALSE--------------------------------------------------------------
-knitr::include_graphics("../man/figures/ggplotly_screenshot.png", dpi = 300)
+    caption = "Edgar Anderson, 1935")
 
 ## -----------------------------------------------------------------------------
 df <- data.frame(
-  trt = factor(c(1, 1, 2, 2)),
-  resp = c(1, 5, 3, 4),
+  treatment = factor(c(1, 1, 2, 2)),
+  response = c(1, 5, 3, 4),
   group = factor(c(1, 2, 1, 2)),
   upper = c(1.1, 5.3, 3.3, 4.2),
   lower = c(0.8, 4.6, 2.4, 3.6)
 )
 
-df %>% 
+df %>%  
   gg_col(
-    x = trt,
-    y = resp,
+    x = response,
+    y = treatment,
     col = group,
     position = "dodge",
-    y_include = max(df$upper) 
-  ) +
+    x_include = max(.$upper),
+    width = 0.66, 
+    pal = c("#1B9E77", "#9E361B")) +
   geom_errorbar(
-    aes(x = trt, ymin = lower, ymax = upper, group = group),
-    inherit.aes = FALSE,
+    aes(xmin = lower, xmax = upper, group = group),
     col = "#7F7F7F",
-    position = position_dodge(width = 0.9),
-    width = 0.1
-    )
+    position = position_dodge(width = 0.66),
+    width = 0.1)
 
 ## -----------------------------------------------------------------------------
-penguins |> 
-  gg_blank(x = flipper_length_mm, y = body_mass_g, col = species) +
-  geom_point(aes(shape = species)) +
-  labs(shape = "Species")
-
-## -----------------------------------------------------------------------------
-penguins %>% 
-  gg_point(
-    x = flipper_length_mm, 
-    y = body_mass_g, 
-    coord = coord_cartesian(xlim = c(190, 200)),
-    x_breaks = scales::breaks_width(2))
-
-## ---- fig.asp=0.75------------------------------------------------------------
-penguins |>
-  tidyr::drop_na() |>
-  mutate(sex = str_to_sentence(sex)) |>
-  gg_jitter(
-    x = sex, 
-    y = body_mass_g,
-    col = sex, 
-    size = 1, 
-    x_title = "",
-    y_limits = c(NA, NA)) +
-  facet_grid(cols = vars(species),
-             rows = vars(island), 
-             switch = "y")
-
-## -----------------------------------------------------------------------------
-diamonds |> 
-  ggplot() +
-  geom_density2d(aes(x = carat, y = price, colour = after_stat(level))) +
-  scale_colour_gradientn(colours = viridis::viridis(9)) +
-  scale_y_continuous(expand = c(0, 0), limits = c(0, 6000)) +
-  labs(x = "Carat", y = "Price", colour = "Level") +
-  gg_theme() +
-  theme(panel.grid.major.x = element_blank()) +
-  theme(legend.position = "right") +
-  theme(legend.direction = "vertical")
+iris |>
+  mutate(Species = stringr::str_to_sentence(Species)) |> 
+  gg_blank(
+    x = Sepal.Width,
+    y = Sepal.Length,
+    col = Species,
+    facet = Species,
+    col_legend_place = "r",
+    col_labels = snakecase::to_sentence_case) +
+  ggdensity::geom_hdr(colour = NA) +
+  labs(alpha = "Probs") +
+  theme(legend.title = element_text(margin = margin(t = 5)))